* All tools:    - Change: Adaptation to cmake 2.8. No more "add_subdirectory()" when        the sub directory doesn't hold a CMakeList.txt.          For "add_definition()", set policy 00005 to NEW.    - Change: Integrate IO extenal libraries.    - Change: Small compatibilities between "size_t" and "unsigned int"        under 64 bits.    - Change: Temporarily disable Equinox/Solstice in unicorn.

+++ katabatic/src/AutoContact.cpp
@@ -2237,7 +2237,7 @@ namespace Katabatic {
 
   unsigned int  AutoContact::getMinDepth () const
   {
-    unsigned int minDepth = (unsigned int)-1;
+    size_t minDepth = (size_t)-1;
     Component* anchor = getAnchor ();
     if ( anchor ) {
       minDepth = min ( minDepth, Session::getRoutingGauge()->getLayerDepth(anchor->getLayer()) );
@@ -2249,7 +2249,7 @@ namespace Katabatic {
     //ltrace(200) << "Slave:" << *icomponent << endl;
     }
 
-    return minDepth;
+    return (unsigned int)minDepth;
   }
